+++ app-backend/db/src/main/scala/MVTLayerDao.scala
@@ -25,6 +25,7 @@ object MVTLayerDao {
       geom: Projected[Geometry],
       envelope: ProjectedExtent,
       taskId: UUID,
+      score: Option[Float],
       labelClassId: UUID,
       className: String,
       colorHexCode: String
@@ -34,8 +35,8 @@ object MVTLayerDao {
         "annotation_task_id" -> VString(taskId.toString),
         "label_class_id" -> VString(labelClassId.toString),
         "name" -> VString(className),
-        "color_hex_code" -> VString(colorHexCode)
-      )
+        "color_hex_code" -> VString(colorHexCode),
+      ) ++ score.fold(Map.empty[String, Value])(v => Map("score" -> VDouble(v)))
   }
 
   private[database] def getAnnotationProjectTasksQ(
@@ -83,6 +84,7 @@ object MVTLayerDao {
           annotation_labels.geometry,
           ST_TileEnvelope(${z}, ${x}, ${y}) as envelope,
           annotation_labels.annotation_task_id,
+          annotation_labels.score,
           annotation_label_classes.id as label_class_id,
           annotation_label_classes.name,
           annotation_label_classes.color_hex_code