+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Included a README in STAC exports for campaigns [#5610](https://github.com/raster-foundry/raster-foundry/pull/5610/)
 - Added a `score` field to labels [#5614](https://github.com/raster-foundry/raster-foundry/pull/5614)
 - Use the presence of groundworkProUser key in verified JWTs to upgrade users' limits [#5615](https://github.com/raster-foundry/raster-foundry/pull/5615)
+- Added human-in-the-loop info to annotation project routes and label MVT [#5617](https://github.com/raster-foundry/raster-foundry/pull/5617)
 
 ### Removed
 - Tore down backsplash-export module and job [#5610](https://github.com/raster-foundry/raster-foundry/pull/5610/)